requirements/base.txt
celery==5.3.6
Django==4.2.7
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-filter==23.4
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
django-ckeditor=6.7.1
drf-yasg==1.21.7
Faker==20.1.0
flower==2.0.1
Jinja2==3.1.2
model-bakery==1.17.0
numpy==1.24.4
Pillow==10.1.0
PyJWT==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
urllib3==2.1.0
psycopg2-binary==2.9.9
async-timeout==4.0.3